task done for gree logix

+++ database/migrations/2022_05_13_220658_create_affiliates_table.php
@@ -18,7 +18,8 @@ return new class extends Migration
             $table->foreignId('user_id');
             $table->foreignId('merchant_id');
             // TODO: Replace me with a brief explanation of why floats aren't the correct data type, and replace with the correct data type.
-            $table->float('commission_rate');
+            // Using decimal instead of float to avoid precision issues
+            $table->decimal('commission_rate', 5, 2);
             $table->string('discount_code');
             $table->timestamps();
         });
